+++ src/ledger_formatter.rs
@@ -1,12 +1,12 @@
 /*!
  * Formats prices for Ledger
  */
+use as_symbols::SymbolMetadata;
 
-use crate::model::PriceWSymbol;
+use crate::model::Price;
 
-/*
 /// format a list of prices
-pub(crate) fn format_prices(prices: Vec<Price>, securities: &[Security]) -> String {
+pub(crate) fn format_prices(prices: Vec<Price>, securities: Vec<SymbolMetadata>) -> String {
     let mut output = String::default();
 
     for price in prices {
@@ -16,95 +16,91 @@ pub(crate) fn format_prices(prices: Vec<Price>, securities: &[Security]) -> Stri
         // find the matching security by id
         let sec = securities
             .iter()
-            .find(|x| x.id == price.security_id)
+            .find(|sm| sm.symbol_w_namespace() == price.symbol)
             .expect("a matching security");
-        // let sec = securities[&price.security_id];
+        // let sec = securities[&price.symbol];
 
         output += format_price(&price, sec).as_str();
         output += "\n";
     }
     output
 }
-*/
 
-pub fn format_prices_w_symbols(prices: Vec<PriceWSymbol>) -> String {
-    let mut output = String::default();
-    for pws in prices {
-        output += format_price_w_symbol(&pws).as_str();
-        output += "\n";
-    }
-    output
-}
+// pub fn format_prices_w_symbols(prices: Vec<Price>) -> String {
+//     let mut output = String::default();
+//     for pws in prices {
+//         output += format_price_w_symbol(&pws).as_str();
+//         output += "\n";
+//     }
+//     output
+// }
 
 /** Formats a single Price record.
  * ledger price format, ISO format supported:
  * P 2004-06-21 02:17:58 VTI $27.76
  */
-/*
-fn format_price(price: &Price, sec: &Security) -> String {
+fn format_price(price: &Price, sec: &SymbolMetadata) -> String {
     let date = price.date.to_owned();
     let time = price.time.to_owned();
     let date_time = format!("{date} {time}");
 
-    let mnemonic = match sec.ledger_symbol {
-        Some(_) => sec.ledger_symbol.to_owned(),
+    let symbol = match &sec.ledger_symbol {
+        Some(ledger_symbol) => Some(ledger_symbol.to_owned()),
         None => Some(sec.symbol.to_owned()),
     }
     .expect("valid symbol");
     let value = price.to_decimal();
     let currency = &price.currency;
 
-    format!("P {date_time} {mnemonic} {value} {currency}")
-}
- */
-
-fn format_price_w_symbol(pws: &PriceWSymbol) -> String {
-    let date = pws.date.to_owned();
-    let time = pws.time.to_owned();
-    let date_time = format!("{date} {time}");
-
-    let symbol = if pws.ledger_symbol.is_empty() {
-        &pws.symbol
-    } else {
-        &pws.ledger_symbol
-    };
-    let value = pws.value;
-    let currency = &pws.currency;
-
     format!("P {date_time} {symbol} {value} {currency}")
 }
 
+// fn format_price_w_symbol(pws: &Price) -> String {
+//     let date = pws.date.to_owned();
+//     let time = pws.time.to_owned();
+//     let date_time = format!("{date} {time}");
+
+//     let symbol = if pws.ledger_symbol.is_empty() {
+//         &pws.symbol
+//     } else {
+//         &pws.ledger_symbol
+//     };
+//     let value = pws.value;
+//     let currency = &pws.currency;
+
+//     format!("P {date_time} {symbol} {value} {currency}")
+// }
+
 #[cfg(test)]
 mod tests {
-    
 
     // use super::*;
 
     /*
-    #[test]
-    fn test_single_price_formatting() {
-        let mut symbol = Security::new();
-        symbol.ledger_symbol = Some("EL4X_DE".into());
+       #[test]
+       fn test_single_price_formatting() {
+           let mut symbol = Security::new();
+           symbol.ledger_symbol = Some("EL4X_DE".into());
 
-        let price = Price {
-            id: 113,
-            security_id: 26,
-            date: "2022-12-01".into(),
-            time: "12:25:34".into(),
-            value: 12534,
-            denom: 100,
-            currency: "EUR".into(),
-        };
+           let price = Price {
+               id: 113,
+               security_id: 26,
+               date: "2022-12-01".into(),
+               time: "12:25:34".into(),
+               value: 12534,
+               denom: 100,
+               currency: "EUR".into(),
+           };
 
-        let actual = format_price(&price, &symbol);
+           let actual = format_price(&price, &symbol);
 
-        // println!("{actual:?}");
+           // println!("{actual:?}");
 
-        assert_eq!(actual, "P 2022-12-01 12:25:34 EL4X_DE 125.34 EUR");
-    }
- */
+           assert_eq!(actual, "P 2022-12-01 12:25:34 EL4X_DE 125.34 EUR");
+       }
+    */
 
- /*
+    /*
     #[test]
     fn test_format_price_wo_time() {
         let mut symbol = Security::new();